minor clean up of prom files
