--- pkg/web/web.go
@@ -1,24 +1,0 @@
-package web
-
-import (
-	"net/http"
-
-	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/sirupsen/logrus"
-)
-
-// Listen start listening http
-func Listen(addr string) *http.Server {
-	mux := http.NewServeMux()
-	mux.Handle("/metrics", promhttp.Handler())
-	srv := http.Server{
-		Addr:    addr,
-		Handler: mux,
-	}
-	go func() {
-		if err := srv.ListenAndServe(); err != nil {
-			logrus.Fatal(err)
-		}
-	}()
-	return &srv
-}
